requirements.txt
streamlit
Pillow
opencv-python
tempfile